Add const in parameters and set time variables as unsigned long

+++ lib/Task/Task.cpp
@@ -1,12 +1,12 @@
 #include "Task.h"
 #include "Component.h"
 
-void Task::init(const int period) {
+void Task::init(const unsigned long period) {
   myPeriod = period;
   timeElapsed = 0;
 }
 
-bool Task::updateAndCheckTime(const int basePeriod) {
+bool Task::updateAndCheckTime(const unsigned long basePeriod) {
   timeElapsed += basePeriod;
   if(timeElapsed >= myPeriod) {
     timeElapsed = 0;
